# docs/openapi/compute-service.yaml
...
components:
  schemas:
    Corev1ConfigMapEnvSource:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      type: object
    Corev1ConfigMapKeySelector:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        key:
          type: string
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      - key
      type: object
    Corev1EnvFromSource:
      properties:
        configMapRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ConfigMapEnvSource'
          nullable: true
        prefix:
          type: string
        secretRef:
          allOf:
          - $ref: '#/components/schemas/Corev1SecretEnvSource'
          nullable: true
      type: object
    Corev1EnvVar:
      properties:
        name:
          type: string
        value:
          type: string
        valueFrom:
          allOf:
          - $ref: '#/components/schemas/Corev1EnvVarSource'
          nullable: true
      required:
      - name
      type: object
    Corev1EnvVarSource:
      properties:
        configMapKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ConfigMapKeySelector'
          nullable: true
        fieldRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ObjectFieldSelector'
          nullable: true
        fileKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1FileKeySelector'
          nullable: true
        resourceFieldRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ResourceFieldSelector'
          nullable: true
        secretKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1SecretKeySelector'
          nullable: true
      type: object
    Corev1FileKeySelector:
      properties:
        key:
          type: string
        optional:
          nullable: true
          type: boolean
        path:
          type: string
        volumeName:
          type: string
      required:
      - key
      - path
      - volumeName
      type: object
    Corev1LocalObjectReference:
      properties:
        name:
          type: string
      type: object
    Corev1ObjectFieldSelector:
      properties:
        apiVersion:
          type: string
        fieldPath:
          type: string
      required:
      - fieldPath
      type: object
    Corev1ResourceClaim:
      properties:
        name:
          type: string
        request:
          type: string
      required:
      - name
      type: object
    Corev1ResourceFieldSelector:
      properties:
        containerName:
          type: string
        divisor:
          description: Kubernetes resource.Quantity (e.g. "500m", "2Gi", "4").
          type: string
        resource:
          type: string
      required:
      - resource
      type: object
    Corev1ResourceRequirements:
      properties:
        claims:
          items:
            $ref: '#/components/schemas/Corev1ResourceClaim'
          type: array
        limits:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
        requests:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
      type: object
    Corev1SecretEnvSource:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      type: object
    Corev1SecretKeySelector:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        key:
          type: string
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      - key
      type: object
    Corev1Toleration:
      properties:
        effect:
          type: string
        key:
          type: string
        operator:
          type: string
        tolerationSeconds:
          format: int64
          nullable: true
          type: integer
        value:
          type: string
      type: object
...
    MLServiceBackend:
      properties:
        config:
          description: Embedded Kubernetes resource (free-form JSON).
          nullable: true
          type: object
        engine:
          type: string
        name:
          type: string
      required:
      - engine
      - name
      type: object
    MLServiceCreateInput:
      properties:
        backend:
          allOf:
          - $ref: '#/components/schemas/MLServiceBackend'
          nullable: true
        description:
          type: string
        displayName:
          type: string
        modelRef:
          $ref: '#/components/schemas/MLServiceModelRef'
        name:
          maxLength: 40
          minLength: 3
          pattern: ^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$
          type: string
        quota:
          type: string
        resourceUnitId:
          format: uuid
          type: string
        roles:
          items:
            $ref: '#/components/schemas/MLServiceRoleSpec'
          minItems: 1
          type: array
        route:
          allOf:
          - $ref: '#/components/schemas/MLServiceRoute'
          nullable: true
        runPolicy:
          allOf:
          - $ref: '#/components/schemas/MLServiceRunPolicy'
          nullable: true
      required:
      - modelRef
      - name
      - quota
      - resourceUnitId
      - roles
      type: object
    MLServiceList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MLServiceView'
          type: array
        total:
          format: int64
          type: integer
      required:
      - items
      - total
      type: object
    MLServiceModelRef:
      properties:
        name:
          type: string
        version:
          type: string
      required:
      - name
      - version
      type: object
    MLServicePodPort:
      properties:
        containerPort:
          format: int32
          type: integer
        name:
          type: string
        protocol:
          type: string
      required:
      - containerPort
      - name
      type: object
    MLServicePodTemplate:
      properties:
        args:
          items:
            type: string
          type: array
        command:
          items:
            type: string
          type: array
        env:
          items:
            $ref: '#/components/schemas/Corev1EnvVar'
          type: array
        envFrom:
          items:
            $ref: '#/components/schemas/Corev1EnvFromSource'
          type: array
        image:
          type: string
        imagePullPolicy:
          type: string
        ports:
          items:
            $ref: '#/components/schemas/MLServicePodPort'
          type: array
        resources:
          $ref: '#/components/schemas/Corev1ResourceRequirements'
        workingDir:
          type: string
      required:
      - image
      type: object
    MLServiceRoleSpec:
      properties:
        name:
          type: string
        replicas:
          format: int32
          type: integer
        template:
          $ref: '#/components/schemas/MLServicePodTemplate'
      required:
      - name
      - replicas
      - template
      type: object
    MLServiceRoute:
      properties:
        auth:
          allOf:
          - $ref: '#/components/schemas/MLServiceRouteAuth'
          nullable: true
        enabled:
          type: boolean
        hostname:
          type: string
        path:
          type: string
        portName:
          type: string
        rateLimit:
          allOf:
          - $ref: '#/components/schemas/MLServiceRouteRateLimit'
          nullable: true
        targetRole:
          type: string
        timeout:
          type: string
      required:
      - enabled
      type: object
    MLServiceRouteAuth:
      properties:
        apiKey:
          allOf:
          - $ref: '#/components/schemas/MLServiceRouteAuthAPIKeyConfig'
          nullable: true
        jwt:
          allOf:
          - $ref: '#/components/schemas/MLServiceRouteAuthJWTConfig'
          nullable: true
        type:
          type: string
      type: object
    MLServiceRouteAuthAPIKeyConfig:
      properties:
        secretRef:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
      required:
      - secretRef
      type: object
    MLServiceRouteAuthJWTConfig:
      properties:
        issuer:
          type: string
        jwksUri:
          type: string
      required:
      - issuer
      - jwksUri
      type: object
    MLServiceRouteRateLimit:
      properties:
        burst:
          format: int32
          type: integer
        requestsPerSecond:
          format: int32
          type: integer
      required:
      - requestsPerSecond
      type: object
    MLServiceRunPolicy:
      properties:
        progressDeadlineSeconds:
          format: int32
          nullable: true
          type: integer
      type: object
    MLServiceScaleInput:
      properties:
        replicas:
          format: int32
          minimum: 0
          type: integer
      required:
      - replicas
      type: object
    MLServiceScheduling:
      properties:
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        priorityClass:
          type: string
        quota:
          type: string
        tolerations:
          items:
            $ref: '#/components/schemas/Corev1Toleration'
          type: array
      required:
      - quota
      type: object
    MLServiceSpec:
      properties:
        backend:
          $ref: '#/components/schemas/MLServiceBackend'
        modelRef:
          $ref: '#/components/schemas/MLServiceModelRef'
        roles:
          items:
            $ref: '#/components/schemas/MLServiceRoleSpec'
          type: array
        route:
          allOf:
          - $ref: '#/components/schemas/MLServiceRoute'
          nullable: true
        runPolicy:
          $ref: '#/components/schemas/MLServiceRunPolicy'
        scheduling:
          $ref: '#/components/schemas/MLServiceScheduling'
      required:
      - backend
      - modelRef
      - roles
      - scheduling
      type: object
    MLServiceView:
      properties:
        createdAt:
          format: date-time
          type: string
        description:
          type: string
        displayName:
          type: string
        endpoint:
          type: string
        id:
          format: uuid
          type: string
        message:
          type: string
        name:
          type: string
        namespace:
          type: string
        ownerUser:
          type: string
        readyReplicas:
          format: int32
          type: integer
        replicas:
          format: int32
          type: integer
        spec:
          $ref: '#/components/schemas/MLServiceSpec'
        status:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - name
      - namespace
      - readyReplicas
      - replicas
      - spec
      - status
      - updatedAt
      type: object
    Problem:
      properties:
        code:
          description: Discrete business error class.
          enum:
          - validation_failed
          - not_found
          - conflict
          - precondition_failed
          - unauthorized
          - forbidden
          - service_unavailable
          - internal_error
          - quota_exceeded
          type: string
        detail:
          type: string
        details:
          additionalProperties: {}
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
...
info:
  description: REST API for resource pools, resource units, jobs, and services. Partitioned
    by Kubernetes namespace. RFC7807 Problem responses on errors.
  title: AxisML Compute Service API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/namespaces/{namespace}/services:
    get:
      operationId: listMLServices
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Page size (1–200, default 50).
        in: query
        name: limit
        schema:
          format: int32
          type: integer
      - description: Number of items to skip.
        in: query
        name: offset
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLServiceList'
          description: Page.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: List services in a namespace
      tags:
      - services
    post:
      operationId: createMLService
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MLServiceCreateInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLServiceView'
          description: Created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Submit an MLService
      tags:
      - services
  /api/v1/namespaces/{namespace}/services/{service}:
    delete:
      operationId: deleteMLService
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Service name.
        in: path
        name: service
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted (idempotent).
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Delete service
      tags:
      - services
    get:
      operationId: getMLService
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Service name.
        in: path
        name: service
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLServiceView'
          description: Service.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Get service
      tags:
      - services
  /api/v1/namespaces/{namespace}/services/{service}/scale:
    post:
      operationId: scaleMLService
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Service name.
        in: path
        name: service
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MLServiceScaleInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLServiceView'
          description: Scaled service.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Scale a service
      tags:
      - services
...
servers:
- description: Same-origin
  url: /
...